add tests, reorganize
